+++ CHANGELOG.md
@@ -83,3 +83,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Provider timeout fixed and Maps are used now to handle subscriptions (#2955)
 - stripHexPrefix fixed (#2989)
 - BatchRequest error handling fixed for callbacks (#2993)
+- ``reconnected`` event and reconnection timeout option added to WebsocketProvider (#2994)